upgrade to ots 1.0.61

+++ src/main/java/edu/harvard/hul/ois/fits/EbuCoreModel.java
@@ -122,10 +122,10 @@ public class EbuCoreModel {
                     // workaround for handling multiple bitrate values that are sometimes returned by mediainfo
                     String[] bitRateParts = dataValue.split(" ");
                     dataValue = bitRateParts[0];
-                    vfmt.setBitRate(Integer.parseInt(dataValue));
+                    vfmt.setBitRate(Long.parseLong(dataValue));
                     break;
                 case bitRateMax:
-                    vfmt.setBitRateMax(Integer.parseInt(dataValue));
+                    vfmt.setBitRateMax(Long.parseLong(dataValue));
                     break;
                 case bitRateMode:
                     vfmt.setBitRateMode(dataValue.toLowerCase());
